feat: extract error handling into a dedicated SocialError component and update social dashboard
