+++ containers/frontend/web/src/features/social/social-variants/index.ts
@@ -1,1 +1,2 @@
 export { UsersList } from './users-list';
+export { SocialError } from './social-error';
