+++ containers/frontend/web/src/features/social/social-variants/users-list.tsx
@@ -11,28 +11,14 @@ import { OnlineButtons } from './online-buttons';
 import { OfflineButtons } from './offline-buttons';
 import { AcceptActionButton } from './accept-action-button';
 import { RejectActionButton } from './reject-action-button';
-import type { SocialErrorCode } from '../social-dashboard';
 
 interface UsersListProps {
   friends: (FriendPublic | FriendshipPublic)[];
   type: 'request' | 'pending' | 'online' | 'offline';
-  error?: SocialErrorCode;
 }
 
-export function UsersList({ friends, type, error }: UsersListProps) {
+export function UsersList({ friends, type }: UsersListProps) {
   const t = useTranslations('features.social.emptyStates');
-  const tErrors = useTranslations('errors');
-
-  if (error) {
-    const errorMessage = tErrors.has(error) ? tErrors(error) : error;
-    return (
-      <Stack align="center" justify="center" className="py-3 px-3 text-center">
-        <Text variant="caption" color="danger">
-          {errorMessage}
-        </Text>
-      </Stack>
-    );
-  }
 
   if (friends.length === 0) {
     return (
